♻️ 동일한 기능을 가진 fetchImage 메서드를 NetworkManager에서 하나로 통일
